+++ examples/User/JsonApi/Document/UsersDocument.php
@@ -56,6 +56,6 @@ class UsersDocument extends AbstractCollectionDocument
      */
     public function getLinks()
     {
-        return Links::createWithoutBaseUri()->setPagination("/users", $this->domainObject);
+        return Links::createWithoutBaseUri()->setPagination("/?path=/users", $this->domainObject);
     }
 }