+++ examples/User/JsonApi/Resource/ContactResourceTransformer.php
@@ -60,7 +60,7 @@ class ContactResourceTransformer extends AbstractResourceTransformer
     {
         return Links::createWithoutBaseUri(
             [
-                "self" => new Link("/contacts/" . $this->getId($contact))
+                "self" => new Link("/?path=/contacts/" . $this->getId($contact))
             ]
         );
     }